refactor(admin): split site editor panels
Move the pure project, writing, newsletter, link, toggle, and preview components out of site-content-editor.tsx so the main editor owns state and save orchestration instead of also carrying every form panel.

Verified with pnpm exec prettier --check apps/admin/src/app/site-content-editor.tsx apps/admin/src/app/site-content-editor-panels.tsx; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; git diff --check -- apps/admin/src/app/site-content-editor.tsx apps/admin/src/app/site-content-editor-panels.tsx; pnpm validate.
